✨ feat(contrato): escala de nivel municipal para Planos de Contingencia
A amostra de Ipatinga (PlanCon 2025/2026) rendeu zero regras corretas com o
contrato anterior. A tabela de limiares (40 valores, deslizamento e
inundacao) estourou o orcamento de tokens do modelo e saiu truncada; a de
umidade do ar virou taxa de chuva em mm, do estado inteiro, com os niveis do
plano convertidos para as cores do POP -- todos os campos errados, com cara
de regra valida.

Contrato:
- escala nivel_municipal, ordinal (n1..n5) com o nome original em
  nivel_rotulo. Nunca convertida para cor: o "Atencao (Amarelo)" de Ipatinga
  comeca em 35 mm/24h, o Amarelo estadual em 60, e os planos nem concordam
  entre si no vocabulario.
- janela_horas fracionaria (15 min = 0.25), positiva.
- sentido acima/abaixo, com chuva/cota/vazao restritas a "acima".

Tabelas: parser deterministico para a forma janela x Nivel 1..k, com o
processo lido do titulo (deslizamento -> geologia, inundacao -> hidrologia) e
o municipio do nome do arquivo do PlanCon. Nao passa pelos sinonimos das
tabelas estaduais, que dariam roxo ao "Critico" que o plano chama de Laranja.
Linha incompleta ou decrescente entre niveis e ignorada inteira. Ipatinga:
40 regras, todas ok, sem GPU.

regra_id passa a incluir janela e dominio: "96h Nivel 1 = 80 mm" e "24h
Nivel 1 = 80 mm" do mesmo chunk colidiam. A fila de revisao passa a agrupar
tambem por entidade, dominio e sentido -- o mesmo numero em dois municipios
nao pode virar uma decisao so.

Prompt: informa o documento, manda escala municipal sem cor e diz que
grandeza fora do vocabulario nao vira regra.

Propostas antigas continuam legiveis (campos novos com o padrao do contrato).
As 7 que o contrato novo recusa -- janela 0, todas inventadas de trechos sem
limiar -- ficam suspeitas no reclassificar em vez de derrubar o lote.

+++ ingestao/validacao.py
@@ -147,7 +147,7 @@ def _motivo_de_implausibilidade(
 GRANDEZAS_DE_TAXA = {"taxa_precipitacao"}
 
 
-def janela_coerente_com_grandeza(grandeza: str, janela_horas: int | None) -> bool:
+def janela_coerente_com_grandeza(grandeza: str, janela_horas: float | None) -> bool:
     """Uma taxa nao tem janela maior que 1h: o "por hora" ja esta na unidade.
 
     Achado real: o modelo leu "60 mm em 24 horas" como taxa_precipitacao de
